A few improvements to ML Workbench.
1. Support local analysis in trainer so we can match prediction API experience (REST API for training).

2. Support early stop in trainer rather than max-steps=5000 as default.
